+++ inc_internal/message.h
@@ -1,22 +1,21 @@
-// Copyright (c) 2023.  NetFoundry Inc.
+// Copyright (c) 2023-2026.  NetFoundry Inc
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// 	Licensed under the Apache License, Version 2.0 (the "License");
+// 	you may not use this file except in compliance with the License.
+// 	You may obtain a copy of the License at
 //
-// https://www.apache.org/licenses/LICENSE-2.0
+// 	https://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// 	Unless required by applicable law or agreed to in writing, software
+// 	distributed under the License is distributed on an "AS IS" BASIS,
+// 	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// 	See the License for the specific language governing permissions and
+// 	limitations under the License.
 
 #ifndef ZITI_SDK_MESSAGE_H
 #define ZITI_SDK_MESSAGE_H
 
 #include "pool.h"
-#include "edge_protocol.h"
 
 #include <stdlib.h>
 #include <stdint.h>
@@ -108,7 +107,7 @@ message *message_new(pool_t *pool, uint32_t content, const hdr_t *headers, int n
 
 void message_set_seq(message *m, uint32_t *seq);
 
-message* new_inspect_result(uint32_t req_seq, uint32_t conn_id, connection_type_t type, const char *msg, size_t msglen);
+message* new_inspect_result(uint32_t req_seq, uint32_t conn_id, uint8_t type, const char *msg, size_t msglen);
 
 #ifdef __cplusplus
 };
